transactions can be added and listed but need to look at the modifyin…

+++ budget-manager/src/main/java/com/budgetmanager/budget_manager/repository/BudgetRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface BudgetRepository extends JpaRepository<Budget, Long> {
 
     // Custom query to find budgets by user ID
-    List<Budget> findAllByUser_UserId(Long userId);
+    List<Budget> findAllByUser_UserId(int userId);
 
 }
